pyproject.toml
[tool.poetry]
name = "decibelduck-be"
version = "0.1.0"
description = "DecibelDuck API Backend"
authors = ["Cory Dodt <corydodt@gmail.com>", "Ashley Fisher <fish.ash@gmail.com>"]
packages = [
    { include = "decibelduck", from = "src" },
]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.70.0"
uvicorn = "^0.15.0"
peewee-async = "^0.7.2"
aiopg = "^1.3.3"
python-json-logger = "^2.0.2"
gunicorn = "^20.1.0"
google-cloud-storage = "^1.43.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
isort = "^5.10.1"
black = "^21.11b1"
mypy = "^0.910"
tox = "^3.24.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
